OSIS-6317 move search learning unit form
