+++ preparation_inscription/views/ajouter_unites_enseignement.py
@@ -24,7 +24,6 @@
 ##############################################################################
 from typing import List
 
-from django import forms
 from django.shortcuts import redirect
 from django.urls import reverse
 from django.utils.functional import cached_property
@@ -38,35 +37,14 @@ from base.views.common import display_error_messages, display_success_messages
 from ddd.logic.learning_unit.commands import LearningUnitSearchCommand
 from ddd.logic.learning_unit.dtos import LearningUnitSearchDTO
 from ddd.logic.preparation_programme_annuel_etudiant.commands import AjouterUEAuProgrammeCommand
-from ddd.logic.shared_kernel.academic_year.commands import SearchAcademicYearCommand
-from education_group.forms.fields import UpperCaseCharField
 from infrastructure.messages_bus import message_bus_instance
+from preparation_inscription.forms.search_learning_units import SearchLearningUnitForm
 from preparation_inscription.views.consulter_contenu_groupement import RAFRAICHIR_GROUPEMENT_CONTENANT
-
-
-class SearchLearningUnitForm(forms.Form):
-    annee_academique = forms.ChoiceField(
-        label=_("Anac.").capitalize(),
-        required=False
-    )
-    code = UpperCaseCharField(max_length=15, label=_("Code").capitalize(), required=False)
-    intitule = forms.CharField(max_length=30, label=_("Title").capitalize(), required=False)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.__init_academic_year_field()
-
-    def __init_academic_year_field(self):
-        all_academic_year = message_bus_instance.invoke(
-            SearchAcademicYearCommand()
-        )
-        self.fields['annee_academique'].choices = [(ac_year.year, str(ac_year)) for ac_year in all_academic_year]
 
 
 class AjouterUnitesEnseignementView(LoginRequiredMixin, HtmxMixin, TemplateView):
     name = 'ajouter_unites_enseignement_view'
 
-    # FormView
     template_name = "preparation_inscription/ajouter_unites_enseignement.html"
     htmx_template_name = "preparation_inscription/ajouter_unites_enseignement.html"
